requirements.txt
streamlit
google-cloud-vertex-ai
google-cloud-aiplatform
google-adk
google-genai
python-dotenv
